requirements.txt
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
mypy-extensions==0.4.3
numpy==1.20.1
pandas==1.2.2
pathspec==0.8.1
PySocks==1.7.1
python-dateutil==2.8.1
python-gitlab==2.5.0
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
